fix: improve data fetcher response serialisation (#216)
# What ❔

Get rid of the separate interceptor that transforms `BigNumber` to
string, accomplish the same result overriding `BigNumber` `toJSON`
function.

## Why ❔

Separate interceptor is not optimal especially for blocks with lots of
related data, which we btw expect to see more often with higher TPS.
There are blocks where all related info weights 2Mb+ and it takes ~200ms
to transform the whole response object by the interceptor. Since after
the transformation JSON serialisation happens anyway, we can just
override `toJSON` for `BigNumber` and it will be picked up during JSON
serialisation. I checked that with this change generated response is the
same.

Performance improvement for large payloads (2Mb+):
1. Serialisation with interceptor (interceptor + json serialiser): avg
290ms
2. Serialisation with overridden `toJSON` (json serialiser): avg 34ms

But most important event loop is not blocked for a long time.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [X] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [X] Tests for the changes have been added / updated.