fix: configure ethers batching (#301)
# What ❔

Add a configuration for ethers batching. Set the default batching config
in case it's not defined.

## Why ❔

In ethers v6 batching is enabled by default and default values don't
work well in our case. That's why we need a way to configure ethers
batching and decrease the batching throughput by default.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [X] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [X] Tests for the changes have been added / updated.