fix: rename tx created field to tx received at (#343)
# What ❔

Rename tx created field to tx received at, fix the tooltip for the
field.

## Why ❔

The new name better reflects the purpose of the value. As for the
tooltip, it was misleading, as it stated that the displayed time was
when the transaction was added to the block, which is incorrect. The
datetime shown is actually the time when the transaction was received.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [X] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [X] Tests for the changes have been added / updated.