feat: add system contracts on the first run (#372)
# What ❔

This PR addresses an issue where system contracts were not displayed as
contracts in the UI until they were updated. The implemented solution
ensures that system contracts are loaded and saved in the database
immediately after work begins. This guarantees their proper display in
the UI, providing a consistent and accurate reflection of the system's
state.

## Why ❔

System contracts are not displayed as contracts in the UI until they
have been updated.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

This PR fixes: https://github.com/matter-labs/block-explorer/issues/358

- [+] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [+] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.

Co-authored-by: Vasyl Ivanchuk <vasyl.ivanchuk@gmail.com>