fix: app network selection (#492)
# What ❔

Fix the app to correctly select a network when the previously selected
one has been removed or renamed.

## Why ❔

If a network in the config is removed or renamed and this network was
previously selected by the user, it breaks the network selection, making
it impossible to change the network. This scenario should be handled the
same way as if no network had been selected.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [X] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.