fix: update ghcr path to isolate with the repo (#538)
# What ❔

Update the `ghcr.io` path to include the `block-explorer` repo sub-path

## Why ❔

Without this path, it's attempting to generate an org-level package
(which it doesn't have access to do so). Now the package will be created
only within the scope of the repo, `block-explorer`.

This is what we did with `zksync-prividium`, however, we were forced to
do so as we needed to have enforced permissions as that's a private repo
and wanted to keep the associated packages (docker images) also private.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] ~~Tests for the changes have been added / updated.~~
- [ ] ~~Documentation comments have been added / updated.~~